// src/generated/live/LiveWorkspace.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Account } from "./Account";
import type { ClientStatus } from "./ClientStatus";
import type { LiveUiState } from "./LiveUiState";
import type { ObservedMatch } from "./ObservedMatch";
import type { PlaySession } from "./PlaySession";

export type LiveWorkspace = { status: ClientStatus, accounts: Array<Account>, matches: Array<ObservedMatch>, sessions: Array<PlaySession>, ui: LiveUiState, uiRevision: number, idleMinutes: number, };